위젯 분리 작업
